feat(immigration-game): Move nextState() method into the Game abstract class.
The only important abstract function to implement is cellNextState. Specific to each game.
