+++ .gitignore
@@ -2,3 +2,6 @@ bin/
 out/
 setup/
 doc/
+doc_gui/
+settings.json
+target/
